\name{wheatSumStats}
\alias{wheatSumStats}
\title{wheat summary statistics}
\description{

Information from a collection of 599 historical CIMMYT wheat lines. The wheat data set is from 
CIMMYT's Global Wheat Program. Historically, this program has conducted numerous international 
trials across a wide variety of wheat-producing environments. The environments represented in 
these trials were grouped into four basic target sets of environments comprising four 
main agroclimatic regions previously defined and widely used by CIMMYT's Global Wheat Breeding Program. 
The phenotypic trait considered here was the average grain yield (GY) of the 599 wheat lines evaluated 
in each of these four mega-environments. 

Wheat lines were recently genotyped using 1447 Diversity Array Technology (DArT) generated by 
Triticarte Pty. Ltd. (Canberra, Australia; \url{https://www.diversityarrays.com}). The DArT markers 
may take on two values, denoted by their presence or absence. Markers with a minor allele frequency 
lower than 0.05 were removed, and missing genotypes were imputed with samples from the marginal 
distribution of marker genotypes, that is, \eqn{x_{ij}=Bernoulli(\hat p_j)}, where  \eqn{\hat p_j}  
is the estimated allele frequency computed from the non-missing genotypes. The number of DArT 
MMs after edition was 1279.

We split the data set into two subgroups (source group - 346 samples, target group - 253 samples), 
serving the illustration of the three TL-PGS methods.

We estimated prior effects from the source group using a Bayesian shrinkage estimation method 
(a Bayesian model with a Gaussian prior centered at zero, model ‘BRR’ in the BGLR R-package). 
  
We split the target group into (i) a training set (60%), (ii) a calibration set (20%), and (iii) a 
testing set (20%). We conducted GWAS and derived LD matrix using the training set.

}

\usage{
  data(wheatSumStats)
}

\format{

\item{wheat_PRIOR}{A vector contains prior estimates.} 
\item{wheat_LD}{A dgCMatrix contains LD matrix.}
\item{wheat_GWAS}{A data.frame contains GWAS result, with five columns: id (variant ID), beta (variant effects), se (variant standard errors), n (sample sizes for GWAS), allele_freq (variant allele frequency), and rownames as id.}
\item{wheat_VLD.X}{A matrix contains validation genetypes.}
\item{wheat_VLD.y}{A vector contains validation phenotypes.} 
\item{wheat_TST.X}{A matrix contains testing genetypes.}
\item{wheat_TST.y}{A vector contains testing phenotypes.} 

}

\source{
  International Maize and Wheat Improvement Center (CIMMYT), Mexico.
}

\references{
McLaren, C. G., L. Ramos, C. Lopez, and W. Eusebio. 2000. ``Applications of the geneaology manegment system.'' 
In \emph{International Crop Information System. Technical  Development Manual, version VI}, edited by McLaren, C. G., J.W. White 
and P.N. Fox. pp. 5.8-5.13. CIMMyT, Mexico: CIMMyT and IRRI. 

McLaren, C. G., R. Bruskiewich, A.M. Portugal, and A.B. Cosico. 2005. The International Rice Information System. 
A platform for meta-analysis of rice crop data. \emph{Plant Physiology} \bold{139}: 637-642.
}
\keyword{datasets}
