\name{wheatSumStats}
\alias{wheatSumStats}
\title{wheat summary statistics}
\description{

Information from a collection of 599 historical CIMMYT wheat lines. The wheat data set is from 
CIMMYT's Global Wheat Program. Historically, this program has conducted numerous international 
trials across a wide variety of wheat-producing environments.

Wheat lines were recently genotyped using 1447 Diversity Array Technology (DArT) generated by 
Triticarte Pty. Ltd. (Canberra, Australia; \url{https://www.diversityarrays.com}). The DArT markers 
may take on two values, denoted by their presence or absence. Markers with a minor allele frequency 
lower than 0.05 were removed, and missing genotypes were imputed with samples from the marginal 
distribution of marker genotypes, that is, \eqn{x_{ij}=Bernoulli(\hat p_j)}, where  \eqn{\hat p_j}  
is the estimated allele frequency computed from the non-missing genotypes. The number of DArT 
MMs after edition was 1279.

We split the data set into two subgroups (source group - 346 samples, target group - 253 samples), 
serving the illustration of the three TL-PGS methods.

We estimated prior effects from the source group using a Bayesian shrinkage estimation method 
(a Bayesian model with a Gaussian prior centered at zero, model ‘BRR’ in the BGLR R-package). 
  
We split the target group into (i) a training set (60/100), (ii) a calibration set (20/100), and (iii) a 
testing set (20/100). We conducted GWAS and derived LD matrix using the training set.

}

\usage{
  data(wheatSumStats)
}

\format{

The vector wheat_PRIOR contains prior estimates.

The dgCMatrix wheat_LD contains LD matrix.

The data.frame wheat_GWAS contains GWAS result, with five columns: id (variant ID), beta (variant 
effects), se (variant standard errors), n (sample sizes for GWAS), allele_freq (variant allele 
frequency), and rownames as id.

The matrix wheat_VLD.X contains validation genetypes.

The vector wheat_VLD.y contains validation phenotypes.

The matrix wheat_TST.X contains testing genetypes.

The vector wheat_TST.y contains testing phenotypes.

}

\source{
  International Maize and Wheat Improvement Center (CIMMYT), Mexico.
}

\references{
McLaren, C. G., L. Ramos, C. Lopez, and W. Eusebio. 2000. ``Applications of the geneaology manegment system.'' 
In \emph{International Crop Information System. Technical  Development Manual, version VI}, edited by McLaren, C. G., J.W. White 
and P.N. Fox. pp. 5.8-5.13. CIMMyT, Mexico: CIMMyT and IRRI. 

McLaren, C. G., R. Bruskiewich, A.M. Portugal, and A.B. Cosico. 2005. The International Rice Information System. 
A platform for meta-analysis of rice crop data. \emph{Plant Physiology} \bold{139}: 637-642.
}
\keyword{datasets}
